Q2/chatbot-project/pyproject.toml
[project]
name = "chatbot-app"
version = "0.1.0"
description = "A chatbot app using OpenAI agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.15",
    "python-dotenv>=1.1.0",
]

[tool.setuptools]
py-modules = ["chatbot", "main", "my_secrets"]
